Adds modules for remote and internal dest driver; removes unused code; org cleanup
