+++ DeepTest.sln
@@ -1,14 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.Example.EchoChatServer", "Test\Example\Test.Example.EchoChatServer\Test.Example.EchoChatServer.csproj", "{B27C9E68-68B1-46A5-9D12-C9F578D99B0C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeepTest", "DeepTest\DeepTest.csproj", "{175F8A87-38F8-469B-8C1F-B0F09BC53A03}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.Example.EchoChatDTSuite", "Test\Example\Test.Example.EchoChatDTSuite\Test.Example.EchoChatDTSuite.csproj", "{A7FB3487-D07A-461E-B238-38142E01634B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeepTestPlugin", "DeepTestPlugin\DeepTestPlugin.csproj", "{2A68466B-6EF5-4919-99D8-6A6E34B066A4}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NFBenchImport.Services.ClientApplication", "NFBenchImport.Services.ClientApplication\NFBenchImport.Services.ClientApplication.csproj", "{1AA30BDA-C824-4382-B047-5073E344319A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NFBenchImport.Benchmark.Reference", "NFBenchImport.Benchmark.Reference\NFBenchImport.Benchmark.Reference.csproj", "{3977C80A-B8A9-4918-B775-1325052366FF}"
@@ -17,16 +9,24 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.NFBenchImport", "Test.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NFBenchImport.Benchmark.Performance", "NFBenchImport.Benchmark.Performance\NFBenchImport.Benchmark.Performance.csproj", "{967F28A5-7A88-41AA-9401-BC20AA11E314}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DeepTest", "DeepTest", "{4034F715-C916-45DA-BDA0-2C2C84A9D9E9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InternalTestDriver", "DeepTest\InternalTestDriver\InternalTestDriver.csproj", "{7CCC0238-09A9-4B3B-BF6C-989B50C03539}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RemoteTestDriverHandler", "DeepTest\RemoteTestDriverHandler\RemoteTestDriverHandler.csproj", "{23A1F946-550C-42EC-A944-66A62C3BC6B4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RemoteAssertionMessages", "DeepTest\RemoteAssertionMessages\RemoteAssertionMessages.csproj", "{8A3DF506-6F80-46B0-B08E-B166267A1654}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Framework", "DeepTest\Framework\Framework.csproj", "{E57540BB-F4F5-4975-98E6-7FE5C2693B4B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RemoteTestingWrapper", "DeepTest\RemoteTestingWrapper\RemoteTestingWrapper.csproj", "{6E11FF0E-4B96-4235-A61D-9916579BBCCF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x86 = Debug|x86
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{175F8A87-38F8-469B-8C1F-B0F09BC53A03}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{175F8A87-38F8-469B-8C1F-B0F09BC53A03}.Debug|x86.Build.0 = Debug|Any CPU
-		{175F8A87-38F8-469B-8C1F-B0F09BC53A03}.Release|x86.ActiveCfg = Release|Any CPU
-		{175F8A87-38F8-469B-8C1F-B0F09BC53A03}.Release|x86.Build.0 = Release|Any CPU
 		{1AA30BDA-C824-4382-B047-5073E344319A}.Debug|x86.ActiveCfg = Debug|x86
 		{1AA30BDA-C824-4382-B047-5073E344319A}.Debug|x86.Build.0 = Debug|x86
 		{1AA30BDA-C824-4382-B047-5073E344319A}.Release|x86.ActiveCfg = Release|x86
@@ -35,25 +35,40 @@ Global
 		{1EBA1215-37AC-48AA-A441-DFEDB64766D3}.Debug|x86.Build.0 = Debug|Any CPU
 		{1EBA1215-37AC-48AA-A441-DFEDB64766D3}.Release|x86.ActiveCfg = Release|Any CPU
 		{1EBA1215-37AC-48AA-A441-DFEDB64766D3}.Release|x86.Build.0 = Release|Any CPU
-		{2A68466B-6EF5-4919-99D8-6A6E34B066A4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{2A68466B-6EF5-4919-99D8-6A6E34B066A4}.Debug|x86.Build.0 = Debug|Any CPU
-		{2A68466B-6EF5-4919-99D8-6A6E34B066A4}.Release|x86.ActiveCfg = Release|Any CPU
-		{2A68466B-6EF5-4919-99D8-6A6E34B066A4}.Release|x86.Build.0 = Release|Any CPU
+		{23A1F946-550C-42EC-A944-66A62C3BC6B4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{23A1F946-550C-42EC-A944-66A62C3BC6B4}.Debug|x86.Build.0 = Debug|Any CPU
+		{23A1F946-550C-42EC-A944-66A62C3BC6B4}.Release|x86.ActiveCfg = Release|Any CPU
+		{23A1F946-550C-42EC-A944-66A62C3BC6B4}.Release|x86.Build.0 = Release|Any CPU
 		{3977C80A-B8A9-4918-B775-1325052366FF}.Debug|x86.ActiveCfg = Debug|x86
 		{3977C80A-B8A9-4918-B775-1325052366FF}.Debug|x86.Build.0 = Debug|x86
 		{3977C80A-B8A9-4918-B775-1325052366FF}.Release|x86.ActiveCfg = Release|x86
 		{3977C80A-B8A9-4918-B775-1325052366FF}.Release|x86.Build.0 = Release|x86
+		{6E11FF0E-4B96-4235-A61D-9916579BBCCF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6E11FF0E-4B96-4235-A61D-9916579BBCCF}.Debug|x86.Build.0 = Debug|Any CPU
+		{6E11FF0E-4B96-4235-A61D-9916579BBCCF}.Release|x86.ActiveCfg = Release|Any CPU
+		{6E11FF0E-4B96-4235-A61D-9916579BBCCF}.Release|x86.Build.0 = Release|Any CPU
+		{7CCC0238-09A9-4B3B-BF6C-989B50C03539}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7CCC0238-09A9-4B3B-BF6C-989B50C03539}.Debug|x86.Build.0 = Debug|Any CPU
+		{7CCC0238-09A9-4B3B-BF6C-989B50C03539}.Release|x86.ActiveCfg = Release|Any CPU
+		{7CCC0238-09A9-4B3B-BF6C-989B50C03539}.Release|x86.Build.0 = Release|Any CPU
+		{8A3DF506-6F80-46B0-B08E-B166267A1654}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8A3DF506-6F80-46B0-B08E-B166267A1654}.Debug|x86.Build.0 = Debug|Any CPU
+		{8A3DF506-6F80-46B0-B08E-B166267A1654}.Release|x86.ActiveCfg = Release|Any CPU
+		{8A3DF506-6F80-46B0-B08E-B166267A1654}.Release|x86.Build.0 = Release|Any CPU
 		{967F28A5-7A88-41AA-9401-BC20AA11E314}.Debug|x86.ActiveCfg = Debug|x86
 		{967F28A5-7A88-41AA-9401-BC20AA11E314}.Debug|x86.Build.0 = Debug|x86
 		{967F28A5-7A88-41AA-9401-BC20AA11E314}.Release|x86.ActiveCfg = Release|x86
 		{967F28A5-7A88-41AA-9401-BC20AA11E314}.Release|x86.Build.0 = Release|x86
-		{A7FB3487-D07A-461E-B238-38142E01634B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A7FB3487-D07A-461E-B238-38142E01634B}.Debug|x86.Build.0 = Debug|Any CPU
-		{A7FB3487-D07A-461E-B238-38142E01634B}.Release|x86.ActiveCfg = Release|Any CPU
-		{A7FB3487-D07A-461E-B238-38142E01634B}.Release|x86.Build.0 = Release|Any CPU
-		{B27C9E68-68B1-46A5-9D12-C9F578D99B0C}.Debug|x86.ActiveCfg = Debug|x86
-		{B27C9E68-68B1-46A5-9D12-C9F578D99B0C}.Debug|x86.Build.0 = Debug|x86
-		{B27C9E68-68B1-46A5-9D12-C9F578D99B0C}.Release|x86.ActiveCfg = Release|x86
-		{B27C9E68-68B1-46A5-9D12-C9F578D99B0C}.Release|x86.Build.0 = Release|x86
+		{E57540BB-F4F5-4975-98E6-7FE5C2693B4B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E57540BB-F4F5-4975-98E6-7FE5C2693B4B}.Debug|x86.Build.0 = Debug|Any CPU
+		{E57540BB-F4F5-4975-98E6-7FE5C2693B4B}.Release|x86.ActiveCfg = Release|Any CPU
+		{E57540BB-F4F5-4975-98E6-7FE5C2693B4B}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{7CCC0238-09A9-4B3B-BF6C-989B50C03539} = {4034F715-C916-45DA-BDA0-2C2C84A9D9E9}
+		{23A1F946-550C-42EC-A944-66A62C3BC6B4} = {4034F715-C916-45DA-BDA0-2C2C84A9D9E9}
+		{8A3DF506-6F80-46B0-B08E-B166267A1654} = {4034F715-C916-45DA-BDA0-2C2C84A9D9E9}
+		{E57540BB-F4F5-4975-98E6-7FE5C2693B4B} = {4034F715-C916-45DA-BDA0-2C2C84A9D9E9}
+		{6E11FF0E-4B96-4235-A61D-9916579BBCCF} = {4034F715-C916-45DA-BDA0-2C2C84A9D9E9}
 	EndGlobalSection
 EndGlobal
