--- DeepTest/API/AssertionResult.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace DeepTest
-{
-    public class AssertionResult
-    {
-        public string Value;
-        public int Key;
-
-        public AssertionResult(int k, object v)
-        {
-            Key = k;
-            Value = v.ToString();
-        }
-    }
-}
-
