--- DeepTest/DeepTest.csproj
@@ -1,88 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{175F8A87-38F8-469B-8C1F-B0F09BC53A03}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>DeepTest</RootNamespace>
-    <AssemblyName>DeepTest</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug</OutputPath>
-    <DefineConstants>DEBUG;</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>full</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release</OutputPath>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="Mono.Cecil">
-      <HintPath>..\tools\Mono.Cecil\lib\net45\Mono.Cecil.dll</HintPath>
-    </Reference>
-    <Reference Include="Mono.Cecil.Mdb">
-      <HintPath>..\tools\Mono.Cecil\lib\net45\Mono.Cecil.Mdb.dll</HintPath>
-    </Reference>
-    <Reference Include="Mono.Cecil.Pdb">
-      <HintPath>..\tools\Mono.Cecil\lib\net45\Mono.Cecil.Pdb.dll</HintPath>
-    </Reference>
-    <Reference Include="Mono.Cecil.Rocks">
-      <HintPath>..\tools\Mono.Cecil\lib\net45\Mono.Cecil.Rocks.dll</HintPath>
-    </Reference>
-    <Reference Include="Mono.Cecil.Inject">
-      <HintPath>..\lib\Mono.Cecil.Inject.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="API\DTRuntime.cs" />
-    <Compile Include="API\DTProcess.cs" />
-    <Compile Include="Instrumentation\WeavePoint.cs" />
-    <Compile Include="API\DTNodeDefinition.cs" />
-    <Compile Include="API\DTNodeInstance.cs" />
-    <Compile Include="Instrumentation\WeavingAspectLocation.cs" />
-    <Compile Include="Instrumentation\Helpers\System\Diagnostics\StopwatchHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\TypeDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\PropertyDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\FieldDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\EventDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\MethodDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\ModuleDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\Mono.Cecil\Mono.Cecil.Cil\VariableDefinitionHelper.cs" />
-    <Compile Include="Instrumentation\Helpers\System\Threading\ThreadHelper.cs" />
-    <Compile Include="API\AssertionResult.cs" />
-    <Compile Include="Instrumentation\WeavingHandler.cs" />
-  </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
-  <ItemGroup>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="API\" />
-    <Folder Include="Instrumentation\" />
-    <Folder Include="Instrumentation\Helpers\" />
-    <Folder Include="Instrumentation\Helpers\System\" />
-    <Folder Include="Instrumentation\Helpers\System\Diagnostics\" />
-    <Folder Include="Instrumentation\Helpers\Mono.Cecil\" />
-    <Folder Include="Instrumentation\Helpers\Mono.Cecil\Mono.Cecil.Cil\" />
-    <Folder Include="Instrumentation\Helpers\System\Threading\" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\DeepTestPlugin\DeepTestPlugin.csproj">
-      <Project>{2A68466B-6EF5-4919-99D8-6A6E34B066A4}</Project>
-      <Name>DeepTestPlugin</Name>
-    </ProjectReference>
-  </ItemGroup>
-</Project>
