+++ DeepTest/Framework/API/AssertionResult.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Framework
+{
+    public class AssertionResult
+    {
+        public string Value;
+        public int Key;
+
+        public AssertionResult(int k, object v)
+        {
+            Key = k;
+            Value = v.ToString();
+        }
+    }
+}
+
