--- DeepTest/Instrumentation/Helpers/System/Threading/ThreadHelper.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace DeepTest
-{
-    public class ThreadHelper
-    {
-        public ThreadHelper()
-        {
-        }
-    }
-}
-
