--- DeepTest/Instrumentation/WeavePoint.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.IO;
-using Mono.Cecil;
-
-namespace DeepTest
-{
-	public class WeavePoint
-	{
-        public MethodDefinition wpMethodDefinition { get; }
-
-		public WeavePoint (
-			string parentModule,
-			string parentType,
-			string methodToWatch,
-            MethodDefinition methodDefinition
-        )
-        {
-            wpMethodDefinition = methodDefinition;
-        }
-
-        public override string ToString()
-        {
-            return wpMethodDefinition.FullName;
-        }
-	}
-}
-
