--- DeepTest/Instrumentation/WeavingAssertionHandler.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace DeepTest
-{
-    public static class WeavingAssertionHandler
-    {
-        public static void updateResultEntry(string info)
-        {
-            DTRuntime.
-        }
-    }
-}
-
