+++ DeepTest/InternalTestDriver/MyClass.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace InternalTestDriver
+{
+    public class MyClass
+    {
+        public MyClass()
+        {
+        }
+    }
+}
+
