+++ DeepTest/RemoteAssertionMessages/MyClass.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace RemoteAssertionMessages
+{
+    public class MyClass
+    {
+        public MyClass()
+        {
+        }
+    }
+}
+
