+++ DeepTest/RemoteTestDriverHandler/MyClass.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace RemoteTestDriverHandler
+{
+    public class MyClass
+    {
+        public MyClass()
+        {
+        }
+    }
+}
+
