+++ DeepTest/RemoteTestingWrapper/RemoteTestingWrapper.csproj
@@ -3,10 +3,10 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{1EBA1215-37AC-48AA-A441-DFEDB64766D3}</ProjectGuid>
+    <ProjectGuid>{6E11FF0E-4B96-4235-A61D-9916579BBCCF}</ProjectGuid>
     <OutputType>Library</OutputType>
-    <RootNamespace>Test.NFBenchImport</RootNamespace>
-    <AssemblyName>Test.NFBenchImport</AssemblyName>
+    <RootNamespace>RemoteTestingWrapper</RootNamespace>
+    <AssemblyName>RemoteTestingWrapper</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
@@ -29,28 +29,10 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
-    <Reference Include="nunit.framework">
-      <HintPath>..\tools\NUnit\lib\net45\nunit.framework.dll</HintPath>
-    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestBenchmarkReference.cs" />
-    <Compile Include="TestBenchmarkPerformance.cs" />
+    <Compile Include="WeavingAssertionHandler.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
-  <ItemGroup>
-    <ProjectReference Include="..\NFBenchImport.Benchmark.Reference\NFBenchImport.Benchmark.Reference.csproj">
-      <Project>{3977C80A-B8A9-4918-B775-1325052366FF}</Project>
-      <Name>NFBenchImport.Benchmark.Reference</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\NFBenchImport.Services.ClientApplication\NFBenchImport.Services.ClientApplication.csproj">
-      <Project>{1AA30BDA-C824-4382-B047-5073E344319A}</Project>
-      <Name>NFBenchImport.Services.ClientApplication</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\DeepTest\Framework\Framework.csproj">
-      <Project>{E57540BB-F4F5-4975-98E6-7FE5C2693B4B}</Project>
-      <Name>Framework</Name>
-    </ProjectReference>
-  </ItemGroup>
 </Project>
