--- DeepTest/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" ?>
-<packages>
-  <package id="Mono.Cecil" version="0.9.6.4" targetFramework="net45" />
-</packages>
