--- Test/Example/Test.Example.EchoChatDTSuite/Test.Example.EchoChatDTSuite.csproj
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{A7FB3487-D07A-461E-B238-38142E01634B}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>Test.Example.EchoChatDTSuite</RootNamespace>
-    <AssemblyName>Test.Example.EchoChatDTSuite</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug</OutputPath>
-    <DefineConstants>DEBUG;</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>full</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release</OutputPath>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="nunit.framework">
-      <HintPath>..\..\..\tools\NUnit\lib\net45\nunit.framework.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestBenchmarkSecurity.cs" />
-    <Compile Include="TestBenchmarkReliability.cs" />
-  </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\DeepTest\DeepTest.csproj">
-      <Project>{175F8A87-38F8-469B-8C1F-B0F09BC53A03}</Project>
-      <Name>DeepTest</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Test.Example.EchoChatServer\Test.Example.EchoChatServer.csproj">
-      <Project>{B27C9E68-68B1-46A5-9D12-C9F578D99B0C}</Project>
-      <Name>Test.Example.EchoChatServer</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\DeepTestPlugin\DeepTestPlugin.csproj">
-      <Project>{2A68466B-6EF5-4919-99D8-6A6E34B066A4}</Project>
-      <Name>DeepTestPlugin</Name>
-    </ProjectReference>
-  </ItemGroup>
-</Project>
